PA-712: fixed an unhelpful "feature" that automatically filled-in environment cells when the user left out part of the pattern;
PA-713: Fixed a serious problem with changes to feature overrides only being saved for the last features overridden and not any that were previously overridden.

PA-714: Fixed a problem with changes made to an XY chart not getting saved when the save button was clicked while still in the edit mode in a cell.

PA-716, PA-717: Fixed a problem in which PA's window comes to front in the z-order (but doesn't become active) when the mouse is moved over areas of the PA window that display popups (e.g. uncertainties, errors in XY charts, search result tab "Records" tooltip thing, etc.).

PA-720: Fixed a problem with PA crashing when certain search patterns are entered.

PA-721: Fixed a problem with the histogram hashmark units not being the same between each hash mark.