Modified the build process to include zipping up the training projects; Modified the training projects based on some recent decisions of how to organize training materials.
PA-750: Fixed a problem with PA crashing when choosing an invalid RTF editor.

PA-749: Did more work to get the C and V lists in pattern/chart building side panel to update properly when data source changes (or if changes are a result of ambiguous seq. or experimental trans. changes).

PA-748 and PA-747: Made some changes in prose on the feedback dialog.